Add IConfigurationService and (un)configure commands
Add service that can configure and unconfigure components at the user or
system level, and two commands that map to configure and unconfigure.

Call these new commands from the Windows and Mac (un)installers.

Update the help command output to include the new commands.

+++ src/osx/Installer.Mac/uninstall.sh
@@ -1,22 +1,23 @@
 #!/bin/bash
 
-# Unconfigure GCM
-HELPER="$(git config --system credential.helper)"
-if [ "$HELPER" = "/usr/local/share/gcm-core/git-credential-manager-core" ]
+# Ensure we're running as root
+if [ $(id -u) != "0" ]
 then
-	echo "Resetting credential helper to 'osxkeychain'..."
-	sudo git config --system credential.helper osxkeychain
-else
-	echo "GCM was not configured as the Git credential helper."
+	sudo "$0" "$@"
+	exit $?
 fi
 
-# Remove GCM symlink
+# Unconfigure
+echo "Unconfiguring credential helper..."
+/usr/local/share/gcm-core/git-credential-manager-core unconfigure
+
+# Remove symlink
 if [ -L /usr/local/bin/git-credential-manager-core ]
 then
-	echo "Deleting GCM symlink..."
+	echo "Deleting symlink..."
 	rm /usr/local/bin/git-credential-manager-core
 else
-	echo "No GCM symlink found."
+	echo "No symlink found."
 fi
 
 # Forget package installation/delete receipt
@@ -25,8 +26,8 @@ sudo pkgutil --forget com.microsoft.GitCredentialManager
 # Remove application files
 if [ -d /usr/local/share/gcm-core/ ]
 then
-	echo "Deleting GCM application files..."
+	echo "Deleting application files..."
 	sudo rm -rf /usr/local/share/gcm-core/
 else
-	echo "No GCM application files found."
+	echo "No application files found."
 fi
